# Add sync in moe_align kernel to prevent garbled inference results in …