When deploying the DeepSeek R1 model, there is a probabilistic occurrence of garbled text or no output. Our analysis indicates that a synchronization mechanism needs to be added in the **sgl_moe_align_block_size_kernel**. Specifically, a **__syncthreads**() should be used when initializing shared memory to zero; otherwise, the result obtained by **atomicAdd** may be incorrect, particularly when the shared memory contains dirty data (i.e., not fully zeroed)."